fix(layout): re-arrange buttons
